Change of jdk version 24->21 / addition of modelmapper and mapstruct dependencies / reformulation of the structure of the patient, patient service, serviceImpl, patient controller models and creating patient request dto and response dto / creation of the patientMapper interface for input and output requests / update of the securityconfig format / adding constructors to all models / adjusting nomenclature and configurations with the database in application.properties
The change in structure was to better organize the code by streamlining the controller by returning the dto instead of the specific model directly, adding more lombok functions.

+++ src/main/java/com/amosjuda/Management_of_medical_appointments/service/Impl/PatientServiceImpl.java
@@ -1,5 +1,8 @@
 package com.amosjuda.Management_of_medical_appointments.service.Impl;
 
+import com.amosjuda.Management_of_medical_appointments.config.PatientMapper;
+import com.amosjuda.Management_of_medical_appointments.dtos.PatientRequestDto;
+import com.amosjuda.Management_of_medical_appointments.dtos.PatientResponseDto;
 import com.amosjuda.Management_of_medical_appointments.models.Patient;
 import com.amosjuda.Management_of_medical_appointments.repositories.PatientRepository;
 import com.amosjuda.Management_of_medical_appointments.service.PatientService;
@@ -7,48 +10,58 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 
 @Service
 public class PatientServiceImpl implements PatientService {
+
     private final PatientRepository patientRepository;
+    private final PatientMapper mapper;
 
     @Autowired
-    public PatientServiceImpl(PatientRepository patientRepository) {
+    public PatientServiceImpl(PatientRepository patientRepository, PatientMapper mapper) {
         this.patientRepository = patientRepository;
+        this.mapper = mapper;
     }
 
     @Override
-    public Patient savePatient(Patient patient) {
-        return patientRepository.save(patient);
+    public PatientResponseDto savePatient(PatientRequestDto dto) {
+        Patient patient = mapper.toEntity(dto);
+        return mapper.toDto(patientRepository.save(patient));
     }
 
     @Override
-    public List<Patient> getALLPatients() {
-        return patientRepository.findAll();
+    public List<PatientResponseDto> getALLPatients(){
+        return patientRepository.findAll().stream()
+                .map(mapper::toDto)
+                .toList();
     }
 
     @Override
-    public Optional<Patient> getOnePatientById(UUID id) {
-        return patientRepository.findById(id);
+    public PatientResponseDto getOnePatientById(UUID id) {
+        return patientRepository.findById(id)
+                .map(mapper::toDto)
+                .orElseThrow(() -> new RuntimeException("Patient not found"));
     }
 
     @Override
-    public Patient updatePatient(UUID id, Patient patient) {
-        Optional<Patient> existingPatient = patientRepository.findById(id);
-        if (existingPatient.isPresent()) {
-            Patient patientToUpdate = existingPatient.get();
-            patientToUpdate.setName(patient.getName());
-            patientToUpdate.setEmail(patient.getEmail());
-            return patientRepository.save(patientToUpdate);
-        } else {
-            throw new RuntimeException("Query not found with id: " + id);
-        }
+    public PatientResponseDto updatePatient(UUID id, PatientRequestDto dto) {
+        Patient patient = patientRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Patient not found"));
+
+        patient.setName(dto.getName());
+        patient.setEmail(dto.getEmail());
+        patient.setPhone(dto.getPhone());
+        patient.setBirthdate(dto.getBirthdate());
+
+        return mapper.toDto(patientRepository.save(patient));
     }
 
     @Override
     public void deletePatient(UUID id) {
+        if (!patientRepository.existsById(id)) {
+            throw new RuntimeException("Patient not found with id: " + id);
+        }
         patientRepository.deleteById(id);
     }
 }
